(2fa-settings) ♻️ Reorganize assets
