--- assets/admin/style.css
@@ -1,178 +1,0 @@
-.built-admin {
-    margin: 5px 15px 2px;
-}
-
-.built-logo img {
-    max-width: 150px;
-    margin: 15px 0;
-}
-
-.built-save .button.button-built {
-    background: #e42029;
-    border-color: #e42029;
-}
-
-.built-save .button.button-built:hover {
-    background: #951b20;
-}
-
-.built-field {
-    margin: 0 0 15px 0;
-    display: flex;
-    align-items: center;
-}
-
-.built-label {
-    min-width: 150px;
-    font-weight: bold;
-}
-
-.built-dash-logo img {
-    max-width: 100px;
-    float: left;
-    border-radius: 4px;
-}
-
-.built-dash-head {
-    display: flex;
-    align-items: center;
-    margin-bottom: 10px;
-}
-
-.built-dash-logo {
-    margin-right: 15px;
-}
-
-.built-dash-nav span.built-nav-button {
-    background: #2c3338;
-    display: inline-block;
-    padding: 3.5px 7.5px;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    -webkit-transition: all 0.3s ease;
-    -moz-transition: all 0.3s ease;
-}
-
-.built-dash-nav {
-    border-bottom: 1px solid rgb(255 255 255 / 10%);
-    padding: 0 0 7.5px 0;
-    margin: 0 0 7.5px 0;
-}
-
-.built-dash-nav span.built-nav-button.active,
-.built-dash-nav span.built-nav-button:hover {
-    background: #d5111f;
-}
-
-.built-bar-outer {
-    background: #2c3338;
-    height: 5px;
-    border-radius: 4px;
-}
-
-.built-progress {
-    margin: 0 0 15px;
-}
-
-.built-bar-inner {
-    height: 5px;
-    background: #e42029;
-    border-radius: 4px;
-    width: 0;
-}
-
-.built-bar-status {
-    opacity: 0;
-}
-
-.built-loading svg {
-    width: 20px;
-    height: 20px;
-}
-
-.built-submit {
-    display: flex;
-    align-self: center;
-}
-
-.built-loading {
-    display: flex;
-    align-items: center;
-    margin-left: 10px;
-    opacity: 0;
-}
-
-.built-2fa {
-    text-align: center;
-}
-
-.built-panel-icon {
-    font-size: 60px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.built-panel-icon span {
-    display: flex;
-    width: 100px;
-    height: 100px;
-    align-items: center;
-    justify-content: center;
-    background: #d63638;
-    border-radius: 100%;
-}
-
-.built-panel-title h2 {
-    color: #fff;
-    font-size: clamp(20px,5vw,40px);
-}
-
-.built-panel-inner {
-    background: #2c3338;
-    padding: 30px;
-    border-radius: 6px;
-}
-
-.built-panel.built-2fa {
-    max-width: 600px;
-    width: 100%;
-    margin: 30px auto 0;
-    padding: 30px;
-}
-
-.built-panel-qr img {
-    max-width: 200px;
-    border-radius: 6px;
-    margin: 0 0 15px 0;
-}
-
-.built-panel-code {
-    margin: 0 0 15px 0;
-}
-
-.built-panel-code input {
-    text-align: center;
-    background: #000;
-    color: #fff;
-    border: none;
-    max-width: 200px;
-}
-
-.built-panel-actions a,
-.built-panel-actions button {
-    background: #d63638 !important;
-    border-color: #d63638 !important;
-    width: 200px;
-}
-
-.built-2fa-error-message button {
-    background: #d63638;
-    border-radius: 2px;
-    border: none;
-    color: #fff;
-    display: inline-block;
-    padding: 3px 15px 5px;
-    margin-left: 5px;
-}
