--- assets/dash/dash.css
@@ -1,243 +1,0 @@
-#builtmighty_dashboard_widget a {
-    color: #D4121F;
-}
-
-.built-panel {
-    background: #1d2327;
-    padding: 15px;
-    color: #fff;
-    border-radius: 4px;
-    margin-bottom: 5px;
-}
-
-.built-panel:last-child {
-    margin-bottom: 0 !important;
-}
-
-.built-panel ul {
-    background: #2c3338;
-    padding: 15px;
-    border-radius: 4px;
-}
-
-.built-panel ul li {
-    margin: 0 0 2px 0;
-}
-
-.built-panel ul li code {
-    border-radius: 4px;
-    margin-left: 5px;
-}
-
-.built-flag {
-    background: #D4121F;
-    border-radius: 4px;
-    padding: 3.5px 7px;
-    display: inline-block;
-}
-
-.built-code {
-    display: block;
-    padding: 15px;
-    border-radius: 4px;
-    margin: 5px 0 0 0 !important;
-}
-
-#builtmighty_dashboard_widget a.built-button {
-    background: none;
-    color: #D4121F;
-    text-decoration: none;
-    border-radius: 4px;
-    padding: 3.5px 7.5px;
-    display: inline-block;
-    border: 1px solid #D4121F;
-    transition: all 0.3s ease;
-    -webkit-transition: all 0.3s ease;
-    -moz-transition: all 0.3s ease;
-}
-
-#builtmighty_dashboard_widget a.built-button:hover {
-    background: #D4121F;
-    color: #fff;
-}
-
-.built-issue-field input, 
-.built-issue-field textarea,
-.built-issue-screenshot {
-    background: #2c3338;
-    border: 1px solid #2c3338;
-    color: #fff;
-    display: block;
-    width: 100%;
-    border-radius: 4px;
-}
-
-.built-issue-screenshot {
-    text-align: center;
-    color: rgb(255 255 255 / 35%);
-    padding: 10px 0;
-    cursor: pointer;
-    justify-content: center;
-}
-
-.built-issue-field {
-    margin: 0 0 10px 0;
-    display: flex;
-    align-items: center;
-}
-
-.built-issue-field input:focus, .built-issue-field textarea:focus {
-    box-shadow: none;
-    border: 1px solid rgb(255 255 255 / 10%);
-}
-
-.built-issue-field.built-issue-screenshot img {
-    max-width: 90%;
-    max-height: 275px;
-    border-radius: 4px;
-}
-
-input.button.button-built {
-    background: #e42029;
-    border-color: #e42029;
-    box-shadow: none;
-}
-
-input.button.button-built:hover,
-input.button.button-built:focus {
-    background: #951b20;
-    border-color: #e42029;
-    box-shadow: none;
-}
-
-.built-form {
-    display: none;
-}
-
-.built-form.active {
-    display: block;
-}
-
-.built-form-status {
-    border-radius: 4px;
-    padding: 5px 10px;
-}
-
-.built-form-status p {
-    margin: 0;
-}
-
-.built-form-status.success {
-    background: green;
-}
-
-.built-form-status.error {
-    background: #951b20;
-}
-
-.built-issue-field label, .built-issue-field input {
-    flex: 1;
-}
-
-.built-issue-field label {
-    max-width: 80px;
-}
-
-li.built-jira-issue {
-    display: flex;
-    align-items: center;
-    padding: 5px 0px 5px 0;
-    border-bottom: 1px solid rgb(255 255 255 / 10%);
-}
-
-span.jira-issue-assignee {
-    display: flex;
-    align-items: center;
-}
-
-span.jira-issue-assignee img {
-    border-radius: 100%;
-    margin: 0 5px 0 0;
-    max-width: 20px;
-}
-
-#builtmighty_dashboard_widget li.built-jira-issue a, 
-#builtmighty_dashboard_widget li.built-jira-issue span {
-    flex: 1;
-    color: #fff !important;
-    transition: all 0.3s ease;
-    -webkit-transition: all 0.3s ease;
-    -moz-transition: all 0.3s ease;
-}
-
-#builtmighty_dashboard_widget li.built-jira-issue a:hover {
-    transform: translate(3px, 0px);
-}
-
-span.jira-issue-status {
-    text-align: center;
-    max-width: 50px;
-    margin: 0 15px;
-    background: rgb(255 255 255 / 5%);
-    border-radius: 4px;
-}
-
-span.jira-issue-assignee {
-    max-width: 110px;
-}
-
-li.built-jira-issue a {
-    font-weight: bold;
-}
-
-span.jira-issue-status.is-dev {
-    background: #1c2b41;
-}
-
-span.jira-issue-status.is-done {
-    background: #1c3329;
-}
-
-li.built-jira-issue:last-child {
-    border-bottom: none;
-}
-
-li.built-tasklist form button {
-    border-radius: 6px;
-    border: none;
-    background: #1d2327;
-    height: 24px;
-    width: 26px;
-    margin-right: 10px;
-    cursor: pointer;
-    color: rgb(255 255 255 / 0%);
-    transition: all 0.3s ease;
-    -webkit-transition: all 0.3s ease;
-    -moz-transition: all 0.3s ease;
-}
-
-li.built-tasklist form p {
-    margin-top: 5px;
-    font-size: 12px;
-    color: rgb(255 255 255 / 60%);
-}
-
-li.built-tasklist form label {
-    font-weight: bold;
-}
-
-.built-task-list-actions {
-    display: flex;
-    align-items: center;
-}
-
-li.built-tasklist form button:hover,
-li.built-tasklist form .built-task-complete button {
-    color: #635994;
-}
-
-.built-panel ul li.built-tasklist {
-    border-bottom: 1px solid rgb(255 255 255 / 10%);
-    margin-bottom: 10px;
-    display: block;
-}
