+++ classes/security/class-lockdown-log.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * Lockdown log.
+ * 
+ * Log all lockdown attempts.
+ * 
+ * @since   2.0.0
+ */
+namespace BuiltMightyKit\Security;
+class builtLockdownLog {
+
+}
